NFC: move Vulkan attribute code to VulkanAttributes.cpp
This is in preparation to migrate Vulkan attributes to use AttrDef.
